remove dynamic allocation of HWSprite on particles
not much of a performance boost, but still a slight one, increases FPS from 15 to 17 with ~65k particles visible (in release builds)
