+++ src/rendering/hwrenderer/scene/hw_bsp.cpp
@@ -44,6 +44,8 @@
 #include "hw_vertexbuilder.h"
 #include "hw_walldispatcher.h"
 
+#include "p_effect_internal.h"
+
 #ifdef ARCH_IA32
 #include <immintrin.h>
 #endif // ARCH_IA32
@@ -607,10 +609,8 @@ void HWDrawInfo::RenderParticles(subsector_t *sub, sector_t *front)
 			int clipres = mClipPortal->ClipPoint(sp->PT.Pos.XY());
 			if (clipres == PClip_InFront) continue;
 		}
-		
-		assert(sp->spr);
 
-		sp->spr->ProcessParticle(this, &sp->PT, front, sp);
+		sp->spr.ProcessParticle(this, &sp->PT, front, sp);
 	}
 	for (uint32_t i = 0; i < sub->particles.Size(); i++)
 	{
@@ -621,8 +621,7 @@ void HWDrawInfo::RenderParticles(subsector_t *sub, sector_t *front)
 			if (clipres == PClip_InFront) continue;
 		}
 
-		HWSprite sprite;
-		sprite.ProcessParticle(this, &Level->Particles[i], front, nullptr);
+		Level->ParticleSprites[i].ProcessParticle(this, &Level->Particles[i], front, nullptr);
 	}
 	SetupSprite.Unclock();
 }
